+++ Extra/XCode Templates/VCore (Gateways)/Gateway.xctemplate/___FILEBASENAME___Gateway.swift
@@ -6,6 +6,6 @@ import VCore
 // MARK: - ___VARIABLE_productName___ Gateway
 struct ___VARIABLE_productName___Gateway: ___VARIABLE_productName___GatewayProtocol {
     func fetch(with parameters: ___VARIABLE_productName___Parameters) async throws -> ___VARIABLE_productName___Entity {
-        FIXME()
+        TODO()
     }
 }