+++ Extra/XCode Templates/VCore (SwiftUI)/ContentView (MVVM).xctemplate/___FILEBASENAME___ViewModel.swift
@@ -6,13 +6,16 @@ import VCore
 // MARK: - ___VARIABLE_productName___ View Model
 @Observable
 final class ___VARIABLE_productName___ViewModel {
-    // MARK: Properties
+    // MARK: Properties - Architecture
     @ObservationIgnored private let parameters: ___VARIABLE_productName___Parameters
 
     @ObservationIgnored var navigationStackCoordinator: NavigationStackCoordinator!
     var alertParameters: AlertParameters?
     private(set) var progressViewParameters: ProgressViewParameters?
-    
+
+    // MARK: Properties - ???
+    // ...
+
     // MARK: Initializers
     init(parameters: ___VARIABLE_productName___Parameters) {
         self.parameters = parameters