+++ Extra/XCode Templates/VCore (SwiftUI)/View (Dynamic UI Model).xctemplate/___FILEBASENAME___.swift
@@ -4,10 +4,13 @@ import SwiftUI
 
 // MARK: - ___VARIABLE_productName___
 struct ___VARIABLE_productName___: View {
-    // MARK: Properties
+    // MARK: Properties - Architecture
     private let uiModel: ___VARIABLE_productName___UIModel
     private let parameters: ___VARIABLE_productName___Parameters
-    
+
+    // MARK: Properties - ???
+    // ...
+
     // MARK: Initializers
     init(
         uiModel: ___VARIABLE_productName___UIModel = .init(),