+++ Extra/XCode Templates/VCore (UIKit)/UIView (Dynamic UI Model).xctemplate/___FILEBASENAME___Parameters.swift
@@ -1,8 +1,0 @@
-//  ___FILEHEADER___
-
-import Foundation
-
-// MARK: - ___VARIABLE_productName___ Parameters
-struct ___VARIABLE_productName___Parameters {
-    // ...
-}