+++ Extra/XCode Templates/VCore (UIKit)/UIView (Dynamic UI Model).xctemplate/___FILEBASENAME___UIModel.swift
@@ -1,8 +1,0 @@
-//  ___FILEHEADER___
-
-import UIKit
-
-// MARK: - ___VARIABLE_productName___ UI Model
-struct ___VARIABLE_productName___UIModel {
-    var backgroundColor: UIColor = .systemBackground
-}