+++ Extra/XCode Templates/VCore (UIKit)/UIViewController + UITableView (VIPER).xctemplate/Mock___FILEBASENAME___Interactor.swift
@@ -1,8 +1,0 @@
-//  ___FILEHEADER___
-
-import Foundation
-
-// MARK: - Mock ___VARIABLE_productName___ Interactor
-struct Mock___VARIABLE_productName___Interactor: ___VARIABLE_productName___Interactive {
-    // ...
-}