+++ Extra/XCode Templates/VCore (UIKit)/UIViewController + UITableView (VIPER).xctemplate/___FILEBASENAME___Interactor.swift
@@ -1,8 +1,0 @@
-//  ___FILEHEADER___
-
-import Foundation
-
-// MARK: - ___VARIABLE_productName___ Interactor
-struct ___VARIABLE_productName___Interactor: ___VARIABLE_productName___Interactive {
-    // ...
-}